refactor: rename table_creator
