+++ src/meta-srv/src/lib.rs
@@ -34,7 +34,7 @@ pub mod procedure;
 pub mod pubsub;
 pub mod selector;
 pub mod service;
-pub mod table_creator;
+pub mod table_meta_alloc;
 pub mod table_routes;
 
 pub use crate::error::Result;
