+++ src/meta-srv/src/metasrv/builder.rs
@@ -54,7 +54,7 @@ use crate::service::mailbox::MailboxRef;
 use crate::service::store::cached_kv::{CheckLeader, LeaderCachedKvStore};
 use crate::service::store::kv::{KvBackendAdapter, KvStoreRef, ResettableKvStoreRef};
 use crate::service::store::memory::MemStore;
-use crate::table_creator::MetaSrvTableMetadataAllocator;
+use crate::table_meta_alloc::MetaSrvTableMetadataAllocator;
 
 // TODO(fys): try use derive_builder macro
 pub struct MetaSrvBuilder {
@@ -366,7 +366,7 @@ fn build_ddl_manager(
         },
     ));
 
-    let table_creator = Arc::new(MetaSrvTableMetadataAllocator::new(
+    let table_meta_allocator = Arc::new(MetaSrvTableMetadataAllocator::new(
         selector_ctx.clone(),
         selector.clone(),
         table_id_sequence.clone(),
@@ -377,7 +377,7 @@ fn build_ddl_manager(
         datanode_clients,
         cache_invalidator,
         table_metadata_manager.clone(),
-        table_creator,
+        table_meta_allocator,
     ))
 }
